refactor: extract helm module from kubernetes-deployment (#143)
Final piece of #143 — Finding #3 catch-all decomposition. Now that flux
and secrets are out, extract Helmfile workflows into a dedicated `helm`
module so kubernetes-deployment shrinks to a thin orchestrator over
`kcl` (manifest rendering) and `kubectl` (apply).

New helm/ module
================
- DeployHelmfile — wraps dag.Helm().HelmfileOperation(...)
- DeployMicroservices — loops over comma-separated helmfile refs
- Single dep: helm@v0.110.0

kubernetes-deployment retains
=============================
- DeployKcl (KCL render + kubectl apply)
- ApplyManifests (raw YAML files / URLs via kubectl)
- InstallCustomResourceDefinitions (kustomize sources / URLs via kubectl)

Dep cleanup follows: kubernetes-deployment now declares only kcl +
kubernetes (drops helm; no other module imports it). Stale internal/dagger/
SDK files for previously dropped deps (sops, templating, git, helm,
secrets) pruned.

Module doc updated to describe the thin-orchestrator role and link to
the helm/, flux/, and secrets/ siblings.

Breaking change for callers
===========================

  dagger call -m kubernetes-deployment deploy-helmfile      → dagger call -m helm deploy-helmfile
  dagger call -m kubernetes-deployment deploy-microservices → dagger call -m helm deploy-microservices

All flags carry over unchanged. The remaining kubernetes-deployment
functions (deploy-kcl, apply-manifests, install-custom-resource-definitions)
stay where they are.

Docs
====
- New helm/README.md with apply / destroy / template + multi-helmfile
  examples and a migration table
- kubernetes-deployment/README.md gains a "thin orchestrator" intro and
  a "moved to helm" pointer
- Top-level README.md modules table gains a Helm row; the kubernetes-
  deployment quick example switches to apply-manifests, and a new helm
  deploy-helmfile quick example is added

Verified
========
dagger develop + go build + go vet pass cleanly across helm,
kubernetes-deployment, flux, secrets.
