+++ helm/main.go
@@ -1,0 +1,8 @@
+// Helm module hosts Helmfile-driven deployment workflows. Extracted from
+// kubernetes-deployment in #143 so the catch-all kubernetes-deployment
+// module shrinks to a thin orchestrator over kcl + kubectl. See
+// README.md for usage.
+
+package main
+
+type Helm struct{}
