+++ kubernetes-deployment/main.go
@@ -1,8 +1,9 @@
-// KubernetesDeployment module bundles Helm/Helmfile rendering and apply,
-// raw manifest application, KCL-based deployments, CRD installation, and
-// SOPS-secret helpers used to roll out workloads against a Kubernetes
-// cluster. Flux bootstrap/render/apply/destroy lives in the dedicated
-// `flux` module. See README.md for usage.
+// KubernetesDeployment is a thin orchestrator over `kcl` (manifest
+// rendering) and `kubectl` (apply): KCL-based deployments, raw manifest
+// application, and CRD installation. Helmfile workflows live in the
+// dedicated `helm` module; Flux bootstrap/render/apply/destroy lives in
+// the dedicated `flux` module; SOPS / k8s Secret manifest workflows live
+// in the dedicated `secrets` module. See README.md for usage.
 
 package main
